fix: refactor checks and redirects to consider feature flag
